+++ package.json
@@ -41,7 +41,7 @@
         "@angular/language-service": "^4.3.6",
         "@angular/tsc-wrapped": "^4.3.3",
         "@rdkmaster/icon-font": "^1.0.2",
-        "@rdkmaster/jigsaw": "^1.0.0-beta.9",
+        "@rdkmaster/jigsaw": "^1.0.0-beta.18",
         "@types/chalk": "^0.4.31",
         "@types/fs-extra": "^3.0.1",
         "@types/glob": "^5.0.30",
