+++ src/app/e2e-testee/combo-select/labelField/app.module.ts
@@ -1,5 +1,5 @@
 import {NgModule} from "@angular/core";
-import {JigsawTileSelectModule} from "jigsaw/component/tile-select/tile-select";
+import {JigsawTileSelectModule} from "jigsaw/component/tile/tile";
 import {JigsawComboSelectModule} from "jigsaw/component/combo-select/index";
 import {ComboSelectLabelFieldDemo} from "./app.component";
 import {CommonModule} from "@angular/common";
