+++ src/app/e2e-testee/combo-select/open/app.module.ts
@@ -1,6 +1,6 @@
 import {NgModule} from "@angular/core";
 import {OpenComboSelectDemo} from "./app.component";
-import {JigsawTileSelectModule} from "jigsaw/component/tile-select/tile-select";
+import {JigsawTileSelectModule} from "jigsaw/component/tile/tile";
 import {JigsawButtonModule} from "jigsaw/component/button/button";
 import {JigsawComboSelectModule} from "jigsaw/component/combo-select/index";
 import {CommonModule} from "@angular/common";
