+++ src/app/e2e-testee/range-time/gr/app.module.ts
@@ -1,7 +1,7 @@
 import {NgModule} from "@angular/core";
 import {CommonModule} from "@angular/common";
 import {JigsawRangeTimeModule} from "jigsaw/component/range-time/index";
-import {JigsawTileSelectModule} from "jigsaw/component/tile-select/tile-select";
+import {JigsawTileSelectModule} from "jigsaw/component/tile/tile";
 import {RangeTimeGrComponent} from "./app.component";
 @NgModule({
     declarations: [RangeTimeGrComponent],
