+++ src/app/e2e-testee/tileselect/searchable/app.module.ts
@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { JigsawTileSelectModule } from "jigsaw/component/tile-select/tile-select";
+import { JigsawTileSelectModule } from "jigsaw/component/tile/tile";
 import { TileselectSearchableComponent }  from './app.component';
 import {CommonModule} from "@angular/common";
