+++ src/app/live-demo/list/app.component.html
@@ -3,35 +3,35 @@
     <div class="demo-1 live-demo">
         <h3>菜单</h3>
         <j-list width="200" trackItemBy="title" (selectedItemsChange)="handleSelect($event)">
-            <j-list-item *ngFor="let title of titles" [value]="title">
+            <j-list-option *ngFor="let title of titles" [value]="title">
                 <span j-title>{{title.title}}</span>
                 <div j-sub-title>{{title.subTitle}} <span *ngIf="title.subMenu" class="fa fa-angle-right"></span></div>
-            </j-list-item>
+            </j-list-option>
         </j-list>
         <p class="message">选中的菜单是: <span>{{selectedItems1}}</span></p>
     </div>
     <div class="demo-2 live-demo">
         <h3>带checkbox的选项</h3>
         <j-list width="350" trackItemBy="title" (selectedItemsChange)="handleSelect2($event)">
-            <j-list-item *ngFor="let title of titles" #listItem [value]="title">
+            <j-list-option *ngFor="let title of titles" #listItem [value]="title">
                 <div j-title>
                     <jigsaw-checkbox [(checked)]="listItem.selected">{{title.title}}</jigsaw-checkbox>
                 </div>
                 <div j-sub-title>{{title.subTitle}} <span *ngIf="title.subMenu" class="fa fa-angle-right"></span></div>
-            </j-list-item>
+            </j-list-option>
         </j-list>
         <p class="message">选中的菜单是: <span>{{selectedItems2}}</span></p>
     </div>
     <div class="demo-3 live-demo">
         <h3>带checkbox的列表</h3>
         <j-list [multipleSelect]="true" trackItemBy="name" (selectedItemsChange)="handleSelect3($event)">
-            <j-list-item *ngFor="let goods of goodsList" #listItem [value]="goods">
+            <j-list-option *ngFor="let goods of goodsList" #listItem [value]="goods">
                 <div class="goods-box">
                     <jigsaw-checkbox [(checked)]="listItem.selected"></jigsaw-checkbox>
                     <span class="logo fa fa-{{goods.logo}}"></span>
                     <p>{{goods.desc}}</p>
                 </div>
-            </j-list-item>
+            </j-list-option>
         </j-list>
         <p class="message">选中的菜单是: <span>{{selectedItems3}}</span></p>
     </div>
@@ -41,13 +41,13 @@
                 trackItemBy="name"
                 [selectedItems]="selectedItems4"
                 (selectedItemsChange)="handleSelect4($event)">
-            <j-list-item *ngFor="let goods of goodsList" #listItem [value]="goods">
+            <j-list-option *ngFor="let goods of goodsList" #listItem [value]="goods">
                 <div class="goods-box">
                     <jigsaw-checkbox [(checked)]="listItem.selected"></jigsaw-checkbox>
                     <span class="logo fa fa-{{goods.logo}}"></span>
                     <p>{{goods.desc}}</p>
                 </div>
-            </j-list-item>
+            </j-list-option>
         </j-list>
         <p class="message">选中的菜单是: <span>{{selectedItemsStr4}}</span></p>
     </div>
@@ -57,13 +57,13 @@
         <j-list [multipleSelect]="true"
                 trackItemBy="name"
                 (selectedItemsChange)="handleSelect5($event)">
-            <j-list-item *ngFor="let goods of goodsList" #listItem [value]="goods" [disabled]="goods.disabled">
+            <j-list-option *ngFor="let goods of goodsList" #listItem [value]="goods" [disabled]="goods.disabled">
                 <div class="goods-box">
                     <jigsaw-checkbox [(checked)]="listItem.selected" [disabled]="goods.disabled"></jigsaw-checkbox>
                     <span class="logo fa fa-{{goods.logo}}"></span>
                     <p>{{goods.desc}}</p>
                 </div>
-            </j-list-item>
+            </j-list-option>
         </j-list>
         <p class="message">选中的菜单是: <span>{{selectedItemsStr5}}</span></p>
     </div>
