+++ src/app/live-demo/list/app.module.ts
@@ -2,7 +2,7 @@ import {NgModule} from "@angular/core";
 import {ListBasicDemoComponent} from "./app.component";
 import {JigsawListModule} from "jigsaw/component/list/list";
 import {CommonModule} from "@angular/common";
-import {JigsawCheckBoxModule} from "../../../../jigsaw/component/checkbox/index";
+import {JigsawCheckBoxModule} from "../../../jigsaw/component/checkbox/index";
 
 @NgModule({
     imports: [JigsawListModule, CommonModule, JigsawCheckBoxModule],
