+++ src/app/live-demo/tile/app.component.ts
