--- src/jigsaw/component/list/list-item.html
@@ -1,9 +1,0 @@
-<div class="jigsaw-list-item-title-box">
-    <div class="jigsaw-list-item-title">
-        <ng-content select="[j-title]"></ng-content>
-    </div>
-    <div class="jigsaw-list-item-sub-title">
-        <ng-content select="[j-sub-title]"></ng-content>
-    </div>
-</div>
-<ng-content></ng-content>
