+++ src/jigsaw/component/list/list-option.html
@@ -1,0 +1,9 @@
+<div class="jigsaw-list-option-title-box">
+    <div class="jigsaw-list-option-title">
+        <ng-content select="[j-title],[jigsaw-title]"></ng-content>
+    </div>
+    <div class="jigsaw-list-option-sub-title">
+        <ng-content select="[j-sub-title],[jigsaw-sub-title]"></ng-content>
+    </div>
+</div>
+<ng-content></ng-content>
