+++ src/jigsaw/component/tile/common.ts
@@ -1,24 +1,11 @@
-import {
-    AfterContentInit,
-    ChangeDetectorRef,
-    Component, ContentChildren,
-    EventEmitter,
-    forwardRef,
-    Input,
-    NgModule,
-    OnDestroy,
-    OnInit,
-    Output,
-    QueryList,
-} from '@angular/core';
-import {CommonModule} from '@angular/common';
-import {ControlValueAccessor, FormsModule, NG_VALUE_ACCESSOR} from '@angular/forms'
-import {JigsawInputModule} from '../input/input';
-import {AbstractJigsawComponent} from '../common';
-import {CallbackRemoval, CommonUtils} from '../../core/utils/common-utils';
+import {AbstractJigsawComponent} from "../common";
+import {ControlValueAccessor} from "@angular/forms";
+import {AfterContentInit, ChangeDetectorRef, EventEmitter, Input, OnDestroy, OnInit, Output, QueryList} from "@angular/core";
+import {CallbackRemoval, CommonUtils} from "../../core/utils/common-utils";
 import {ArrayCollection} from "../../core/data/array-collection";
+import {JigsawTileOption} from "./tile";
 
-export class AbstractJigsawGroupComponent extends AbstractJigsawComponent implements ControlValueAccessor, OnInit, AfterContentInit, OnDestroy{
+export class AbstractJigsawGroupComponent extends AbstractJigsawComponent implements ControlValueAccessor, OnInit, AfterContentInit, OnDestroy {
     private _removeRefreshCallback: CallbackRemoval;
 
     //设置对象的标识
@@ -44,7 +31,7 @@ export class AbstractJigsawGroupComponent extends AbstractJigsawComponent implem
     @Output() public selectedItemsChange = new EventEmitter<any[]>();
 
     //获取映射的items
-    protected _items: QueryList<AbstractJigsawItemComponent>;
+    protected _items: QueryList<AbstractJigsawOptionComponent>;
 
     //根据选中的item更新selectedItems
     private _updateSelectItems(itemValue, selected): void {
@@ -96,9 +83,9 @@ export class AbstractJigsawGroupComponent extends AbstractJigsawComponent implem
 
     ngOnInit() {
         super.ngOnInit();
-        if(this.trackItemBy){
+        if (this.trackItemBy) {
             this.trackItemBy = (<string>this.trackItemBy).split(/\s*,\s*/g);
-        }else{
+        } else {
             console.warn('please input trackItemBy attribute in jigsaw-title control')
         }
     }
@@ -126,7 +113,8 @@ export class AbstractJigsawGroupComponent extends AbstractJigsawComponent implem
         }
     }
 
-    private _propagateChange: any = () => {};
+    private _propagateChange: any = () => {
+    };
 
     public writeValue(newValue: any): void {
         if (this._selectedItems === newValue) {
@@ -155,7 +143,7 @@ export class AbstractJigsawGroupComponent extends AbstractJigsawComponent implem
     }
 }
 
-export class AbstractJigsawItemComponent extends AbstractJigsawComponent{
+export class AbstractJigsawOptionComponent extends AbstractJigsawComponent {
     @Input() public value: any;
 
     @Input() public disabled: boolean = false;
@@ -170,71 +158,3 @@ export class AbstractJigsawItemComponent extends AbstractJigsawComponent{
         super();
     }
 }
-
-@Component({
-    selector: 'jigsaw-tile',
-    templateUrl: 'tile-select.html',
-    host: {
-        '[style.width]': 'width',
-        '[style.height]': 'height',
-        '[class.jigsaw-tile]': 'true'
-    },
-    providers: [
-        {provide: NG_VALUE_ACCESSOR, useExisting: forwardRef(() => JigsawTileSelect), multi: true},
-    ]
-})
-export class JigsawTileSelect extends AbstractJigsawGroupComponent{
-
-    @Input() public searchable: boolean = false;
-
-    // 默认多选
-    public multipleSelect: boolean = true;
-
-    @ContentChildren(forwardRef(() => JigsawTileOption))
-    protected _items: QueryList<JigsawTileOption>;
-
-}
-
-@Component({
-    selector: 'jigsaw-tile-option',
-    template: '<ng-content></ng-content>',
-    host: {
-        '[style.width]': 'width',
-        '[style.height]': 'height',
-        '[style.line-height]': 'height',
-        '[class.jigsaw-tile-option]': 'true',
-        '[class.jigsaw-tile-option-active]': 'selected',
-        '[class.jigsaw-tile-option-disabled]': 'disabled',
-        '(click)': '_$handleClick()'
-    }
-})
-export class JigsawTileOption extends AbstractJigsawItemComponent {
-
-    constructor(public changeDetector: ChangeDetectorRef) {
-        super(changeDetector);
-    }
-
-    /**
-     * 点击组件触发
-     * @internal
-     */
-    public _$handleClick(): void {
-        if(!this.disabled){
-            this.selectedChange.emit(this);
-        }
-    }
-
-}
-
-@NgModule({
-    imports: [CommonModule, FormsModule, JigsawInputModule],
-    declarations: [JigsawTileSelect, JigsawTileOption],
-    exports: [JigsawTileSelect, JigsawTileOption]
-})
-export class JigsawTileSelectModule {
-
-}
-
-
-
-
