+++ src/jigsaw/core/theming/all-theme.scss
@@ -28,12 +28,12 @@
 @import "../../component/table/table-head";
 @import "../../component/tabs/tab";
 @import "../../component/tag/tag";
-@import "../../component/tile-select/tile-select";
-@import "../../component/tile-select/tile-option";
+@import "../../component/tile/tile";
+@import "../../component/tile/tile-option";
 @import "../../component/time/time";
 @import "../../component/tooltip/tooltip";
 @import "../../component/list/list";
-@import "../../component/list/list-item";
+@import "../../component/list/list-option";
