+++ src/styles.scss
@@ -1,1 +1,4 @@
 /* You can add global styles to this file, and also import other style files */
+@import "../node_modules/@rdkmaster/jigsaw/theming-api";
+
+//@include setTableBody(#333, 12px , red, blue );
